+++ Pandora/Paradigm/Structure/Interface/Set.hs
@@ -5,3 +5,4 @@ import Pandora.Paradigm.Primary.Object.Boolean (Boolean)
 
 class Set t where
 	member :: Setoid a => a -> t a -> Boolean
+	subset :: Setoid a => t a -> t a -> Boolean