+++ Pandora/Paradigm/Structure/Stack.hs
@@ -10,10 +10,10 @@ import Pandora.Pattern.Functor.Alternative ((<+>))
 import Pandora.Pattern.Functor.Avoidable (empty)
 import Pandora.Pattern.Functor.Pointable (point)
 import Pandora.Pattern.Functor.Extractable (extract)
-import Pandora.Pattern.Functor.Traversable (Traversable)
+import Pandora.Pattern.Functor.Traversable (Traversable ((->>)))
 import Pandora.Pattern.Functor.Bindable ((>>=))
 import Pandora.Pattern.Transformer.Liftable (lift)
-import Pandora.Pattern.Object.Setoid (Setoid ((==)))
+import Pandora.Pattern.Object.Setoid (Setoid ((==), (/=)))
 import Pandora.Pattern.Object.Semigroup (Semigroup ((+)))
 import Pandora.Pattern.Object.Monoid (Monoid (zero))
 import Pandora.Paradigm.Primary.Object.Boolean ((?))
@@ -34,7 +34,7 @@ import Pandora.Paradigm.Structure.Ability.Nonempty (Nonempty)
 import Pandora.Paradigm.Structure.Ability.Zipper (Zipper)
 import Pandora.Paradigm.Structure.Ability.Focusable (Focusable (Focusing, focusing), Location (Head), focus)
 import Pandora.Paradigm.Structure.Ability.Insertable (Insertable (insert))
-import Pandora.Paradigm.Structure.Interface.Set (Set (member))
+import Pandora.Paradigm.Structure.Interface.Set (Set (member, subset))
 
 -- | Linear data structure that serves as a collection of elements
 type Stack = Maybe <:.> Construction Maybe
@@ -61,6 +61,7 @@ instance Insertable Stack where
 
 instance Set Stack where
 	member x = maybe False (True !) . find (Predicate (== x))
+	subset ss s = Nothing /= (ss ->> \x -> find (Predicate (== x)) s)
 
 pop :: Stack ~> Stack
 pop (TU stack) = TU $ stack >>= deconstruct